Move all routes except well-known to /api
