+++ backend/vasp/currencies.py
@@ -9,7 +9,7 @@ from vasp.uma_vasp.currencies import CURRENCIES
 def construct_blueprint(
     currency_service: ICurrencyService,
 ) -> Blueprint:
-    bp = Blueprint("currencies", __name__, url_prefix="/currencies")
+    bp = Blueprint("currencies", __name__, url_prefix="/api/currencies")
 
     @bp.get("/")
     def get_all() -> Response:
