+++ frontend/src/lib/backendUrl.ts
@@ -3,5 +3,7 @@ import isDevelopment from "./isDevelopment";
 
 export const getBackendUrl = () => {
   const backendDomain = getBackendDomain();
-  return isDevelopment ? `http://${backendDomain}` : `https://${backendDomain}`;
+  return isDevelopment
+    ? `http://${backendDomain}/api`
+    : `https://${backendDomain}/api`;
 };
